Remove obsolete debug log dialog
