--- app/src/main/res/drawable/icon_pause.xml
@@ -1,2 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<vector xmlns:android="http://schemas.android.com/apk/res/android" android:width="24dp" android:height="24dp" android:viewportWidth="24" android:viewportHeight="24"><path android:pathData="M6 19h4V5H6v14zm8-14v14h4V5h-4z" android:fillColor="#FFFFFFFF" /></vector>
