--- app/src/main/res/drawable/icon_play.xml
@@ -1,2 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<vector xmlns:android="http://schemas.android.com/apk/res/android" android:width="24dp" android:height="24dp" android:viewportWidth="24" android:viewportHeight="24"><path android:pathData="M10 8.64L15.27 12L10 15.36V8.64M8 5v14l11-7L8 5z" android:fillColor="#FFFFFFFF" /></vector>
