--- app/src/main/res/layout/debug_dialog.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:padding="8dp">
-
-        <com.winlator.cmod.runtime.system.ui.LogView
-            android:id="@+id/LogView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:minHeight="180dp" />
-    </LinearLayout>
-</LinearLayout>
