--- app/src/main/res/layout/debug_toolbar.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal">
-
-    <ImageButton
-        style="@style/ListMenuButton"
-        android:id="@+id/BTClear"
-        android:src="@drawable/icon_remove"
-        android:layout_marginRight="8dp" />
-
-    <ImageButton
-        style="@style/ListMenuButton"
-        android:id="@+id/BTPause"
-        android:src="@drawable/icon_pause"
-        android:layout_marginRight="8dp" />
-
-</LinearLayout>
